User를 LoginUser로 변경, age 추가 및 패키지 이동
