--- src/main/java/support/auth/userdetails/User.java
@@ -1,35 +1,0 @@
-package support.auth.userdetails;
-
-import java.util.List;
-
-public class User implements UserDetails {
-    private String username;
-    private String password;
-    private List<String> authorities;
-
-    public User(String username, String password, List<String> authorities) {
-        this.username = username;
-        this.password = password;
-        this.authorities = authorities;
-    }
-
-    @Override
-    public String getUsername() {
-        return username;
-    }
-
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public List<String> getAuthorities() {
-        return authorities;
-    }
-
-    @Override
-    public boolean checkCredentials(Object credentials) {
-        return password.equals(credentials.toString());
-    }
-}
